requirements.txt
flask
gunicorn
google-generativeai
google-auth
python-dotenv